fix(core): move theme logic to component and use in both navbars
