+++ packages/sanity/src/core/studio/components/navbar/tools/ToolVerticalMenu.tsx
@@ -1,8 +1,9 @@
 import React, {forwardRef, useMemo} from 'react'
-import {Button, Stack} from '@sanity/ui'
+import {Box, Button, Card, Stack} from '@sanity/ui'
 import {PlugIcon} from '@sanity/icons'
 import {startCase} from 'lodash'
 import {Tool} from '../../../../config'
+import {Appearance} from '../appearance/appearance'
 import {ToolLinkProps, ToolLink} from './ToolLink'
 
 interface ToolVerticalMenuProps {
@@ -16,32 +17,36 @@ export function ToolVerticalMenu(props: ToolVerticalMenuProps) {
 
   return useMemo(
     () => (
-      <Stack as="ul" space={[1, 2]}>
-        {tools.map((tool) => {
-          const title = tool?.title || startCase(tool.name) || undefined
+      <>
+        <Stack as="ul" space={[1, 2]}>
+          {tools.map((tool) => {
+            const title = tool?.title || startCase(tool.name) || undefined
 
-          const Link = forwardRef(function Link(
-            linkProps: ToolLinkProps,
-            ref: React.Ref<HTMLAnchorElement>
-          ) {
-            return <ToolLink {...linkProps} ref={ref} name={tool.name} />
-          })
+            const Link = forwardRef(function Link(
+              linkProps: ToolLinkProps,
+              ref: React.Ref<HTMLAnchorElement>
+            ) {
+              return <ToolLink {...linkProps} ref={ref} name={tool.name} />
+            })
 
-          return (
-            <Stack as="li" key={tool.name}>
-              <Button
-                as={Link}
-                icon={tool.icon || PlugIcon}
-                justify="flex-start"
-                mode="bleed"
-                selected={activeToolName === tool.name}
-                tabIndex={isVisible ? 0 : -1}
-                text={title}
-              />
-            </Stack>
-          )
-        })}
-      </Stack>
+            return (
+              <Stack as="li" key={tool.name}>
+                <Button
+                  as={Link}
+                  icon={tool.icon || PlugIcon}
+                  justify="flex-start"
+                  mode="bleed"
+                  selected={activeToolName === tool.name}
+                  tabIndex={isVisible ? 0 : -1}
+                  text={title}
+                />
+              </Stack>
+            )
+          })}
+        </Stack>
+
+        <Appearance />
+      </>
     ),
     [activeToolName, isVisible, tools]
   )
